+++ sources/packages/backend/libs/test-utils/src/index.ts
@@ -13,3 +13,4 @@ export * from "./factories/disbursement-schedule";
 export * from "./factories/disbursement-overaward";
 export * from "./factories/disbursement-value";
 export * from "./factories/msfaa-number";
+export * from "./models/province.model";
