+++ sources/packages/backend/libs/test-utils/src/models/province.model.ts
@@ -1,0 +1,18 @@
+/**
+ * Canadian provinces.
+ */
+export enum Provinces {
+  Alberta = "AB",
+  BritishColumbia = "BC",
+  Manitoba = "MB",
+  NewBrunswick = "NB",
+  NewFoundlandAndLabrador = "NL",
+  NovaScotia = "NS",
+  Ontario = "ON",
+  PrinceEdwardIsland = "PE",
+  Quebec = "QC",
+  Saskatchewan = "SK",
+  Yukon = "YT",
+  NorthernTerritories = "NT",
+  Nunavut = "NU",
+}
